backend/pyproject.toml
[project]
name = "asc-project-entry-exit-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery==5.3.4",
    "django==4.2.7",
    "django-cors-headers==4.3.1",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.3.1",
    "openpyxl==3.1.2",
    "pandas==2.1.3",
    "psycopg2-binary==2.9.9",
    "python-decouple==3.8",
    "redis==5.0.1",
]
